feat: improve navigation responsiveness and upcoming section utility
- replace internal anchor navigation with next/link to avoid full page reloads
- remove redundant header-card API fetches in view page and reuse current view payload
- extract upcoming section builder into app/_lib/upcoming_sections.ts for page export safety
- update upcoming sectioning test to import the shared utility and narrow union types safely
- keep existing upcoming UI style tweaks in globals.css while preserving current spacing rules
