+++ app/globals.css
@@ -918,7 +918,7 @@ textarea.note-input {
 }
 
 .upcoming-day-number {
-  font-size: 36px;
+  font-size: 28px;
   line-height: 1;
   font-weight: 700;
   color: #3e434a;
